Refactoring to make consistent with arm branch
