+++ src/main/java/frc/robot/commands/grabber/GrabCommand.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.grabber;
 
 import frc.robot.Constants;
 import frc.robot.subsystems.Grabber;
