+++ src/main/java/frc/robot/commands/masters/GrabMasterCommand.java
@@ -2,8 +2,8 @@ package frc.robot.commands.masters;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.commands.GrabCommand;
 import frc.robot.commands.ReleaseCommand;
+import frc.robot.commands.grabber.GrabCommand;
 import frc.robot.helpers.ColorDetector;
 import frc.robot.subsystems.Grabber;
